Refactor Home dashboard sections
